ㅇㅇㅇㅇ

+++ src/main/java/com/ll/exam/test2/article/ArticleController.java
@@ -51,12 +51,15 @@ public class ArticleController {
     }
 
 
-//    @PreAuthorize("isAuthenticated()")
+    @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable("id") Integer id) {
+    public String questionDelete(Principal principal, @PathVariable("id") Integer id) {
         Article article = this.articleService.getArticle(id);
+        if (!article.getAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
         this.articleService.delete(article);
-        return "redirect:/article/list";
+        return "redirect:/";
     }
 
     @PreAuthorize("isAuthenticated()")
